requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.1
bcrypt==4.2.1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
databases==0.9.0
ecdsa==0.19.0
fastapi==0.115.7
fastapi-limiter==0.1.6
fastavro==1.10.0
google-api-core==2.24.0
google-auth==2.38.0
googleapis-common-protos==1.66.0
h11==0.14.0
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
numpy==2.0.2
opencensus==0.11.4
opencensus-context==0.1.3
opencensus-ext-azure==1.1.14
pandas==2.2.3
passlib==1.7.4
portalocker==2.10.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2024.2
redis==5.3.0b4
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
